test(grid): refactor general test #3467
